+++ src/cmd/to.rs
@@ -41,6 +41,8 @@ JSON options:
                           [default: 512]
     --nulls               Convert empty string to a null value.
     --omit                Ignore the empty values.
+    --strings <columns>   Force selected columns to be considered as raw strings
+                          instead of integers, floats etc.
 
 NPY options:
     --dtype <type>  Number type to use for the npy conversion. Must be one of \"f32\"
@@ -71,6 +73,7 @@ struct Args {
     flag_sample_size: isize,
     flag_nulls: bool,
     flag_omit: bool,
+    flag_strings: Option<SelectedColumns>,
     flag_dtype: String,
 }
 
@@ -108,7 +111,8 @@ impl Args {
     }
 
     fn convert_to_json(&self) -> CliResult<()> {
-        let mut rdr = self.rconf().reader()?;
+        let rconf = self.rconf();
+        let mut rdr = rconf.reader()?;
         let mut writer = self.wconf().buf_io_writer()?;
 
         let headers = rdr.headers()?.clone();
@@ -118,6 +122,17 @@ impl Args {
             self.sample_size(),
             self.json_empty_mode(),
         );
+
+        if let Some(sel) = &self.flag_strings {
+            let indices = sel.selection(
+                headers.iter().map(|cell| cell.as_bytes()),
+                !rconf.no_headers,
+            )?;
+
+            for index in indices.iter() {
+                inferrence_buffer.set_string(*index);
+            }
+        }
 
         inferrence_buffer.read(&mut rdr)?;
 
@@ -143,7 +158,8 @@ impl Args {
     }
 
     fn convert_to_ndjson(&self) -> CliResult<()> {
-        let mut rdr = self.rconf().reader()?;
+        let rconf = self.rconf();
+        let mut rdr = rconf.reader()?;
         let mut writer = self.wconf().buf_io_writer()?;
 
         let headers = rdr.headers()?.clone();
@@ -153,6 +169,17 @@ impl Args {
             self.sample_size(),
             self.json_empty_mode(),
         );
+
+        if let Some(sel) = &self.flag_strings {
+            let indices = sel.selection(
+                headers.iter().map(|cell| cell.as_bytes()),
+                !rconf.no_headers,
+            )?;
+
+            for index in indices.iter() {
+                inferrence_buffer.set_string(*index);
+            }
+        }
 
         inferrence_buffer.read(&mut rdr)?;